Pulseaudio issue workaround
Pulseaudio needs to be running beforehand and the user must have access to it
to be able to connect to the machine (pairing works without this). To fix this
two options are there, run pulseaudio as a system systemd service or add user
to the group which by default runs pulseaudio (gdm). The latter was used here.