remove fsgroup option from charts
fsgroup is plugin specific, with minikube at least it doesnt do
anything. even if it did, I dont think it is necessary for the
filesystem chmod-ed to be owned by group 568, it will just cause high io
usage. disabling, will add again i guess in the future if an issue is
raised